// src/BookBarter.React/src/api/generated/models/messaging-user-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BookBarter.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface MessagingUserDto
 */
export interface MessagingUserDto {
  /**
   *
   * @type {number}
   * @memberof MessagingUserDto
   */
  id?: number;
  /**
   *
   * @type {string}
   * @memberof MessagingUserDto
   */
  userName?: string | null;
  /**
   *
   * @type {string}
   * @memberof MessagingUserDto
   */
  lastMessage?: string | null;
  /**
   *
   * @type {string}
   * @memberof MessagingUserDto
   */
  lastMessageTime?: string | null;
}