// src/BookBarter.React/src/api/generated/models/message-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BookBarter.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface MessageDto
 */
export interface MessageDto {
  /**
   *
   * @type {number}
   * @memberof MessageDto
   */
  id?: number;
  /**
   *
   * @type {number}
   * @memberof MessageDto
   */
  senderId?: number;
  /**
   *
   * @type {number}
   * @memberof MessageDto
   */
  receiverId?: number;
  /**
   *
   * @type {string}
   * @memberof MessageDto
   */
  senderName?: string | null;
  /**
   *
   * @type {string}
   * @memberof MessageDto
   */
  receiverName?: string | null;
  /**
   *
   * @type {string}
   * @memberof MessageDto
   */
  body?: string | null;
  /**
   *
   * @type {string}
   * @memberof MessageDto
   */
  sentTime?: string;
  /**
   *
   * @type {string}
   * @memberof MessageDto
   */
  editTime?: string | null;
}